refactor: improve type stubs for kivy app by only providing stubs for pyx files and making `Widget` use correct `EventListener` base and fix type errors coming to surface

+++ ubo_app/menu_app/menu_footer.py
@@ -119,7 +119,7 @@ class MenuAppFooter(UboApp):
         def set_value(_: int | None = None) -> None:
             clock.text = now().strftime('%H:%M')
 
-        def update(_: int | None = None) -> None:
+        def update(_: float | None = None) -> None:
             set_value()
             now_ = now()
             Clock.schedule_once(
